requirements.txt
graphviz==0.21
matplotlib==3.10.7
numpy==2.3.4
opencv_contrib_python==4.12.0.88
Pillow==12.0.0
pydicom==3.0.1
scipy==1.16.2
skimage==0.0
tensorboardx==2.6.4
tifffile==2025.5.10
torch==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.67.1
visdom==0.2.4